refactor cli commands into self-contained functions
