+++ src/cli.ts
@@ -1,10 +1,8 @@
-import {Command} from 'commander';
-import {audit, initSchema, loadSchema, scanVars, syncEnvFile, validate} from "./lib";
-import process from "process";
-import fs from "fs";
-import {parse} from "dotenv";
-import {updatePackageJson, detectSourceDirectory} from "./install";
-import path from "path";
+import { Command } from 'commander';
+import { initCommand, InitOptions } from './commands/init';
+import { auditCommand, AuditOptions } from './commands/audit';
+import { validateCommand } from './commands/validate';
+import { syncCommand } from './commands/sync';
 
 export function createCli() {
 
@@ -23,121 +21,30 @@ export function createCli() {
         .option('-f, --force', 'Overwrite existing schema file', false)
         .option('--no-git', 'Disable Git tracking for file scanning')
         .option('--no-scripts', 'Skip adding scripts to package.json')
-        .action((dir, options) => {
-            // Auto-detect source directory if not specified
-            const sourceDir = dir ? (dir.endsWith('/') ? dir : `${dir}/`) : detectSourceDirectory();
-            console.log(`Using source directory: ${sourceDir}`);
-
-            // Update package.json with env-tool scripts if not disabled
-            if (options.scripts) {
-                const packageJsonPath = path.resolve(process.cwd(), 'package.json');
-                if (!fs.existsSync(packageJsonPath)) {
-                    console.error('No package.json found in the current directory.');
-                    process.exitCode = 1;
-                    return;
-                }
-
-                try {
-                    const packageJsonContent = fs.readFileSync(packageJsonPath, 'utf-8');
-                    const packageJson = JSON.parse(packageJsonContent);
-                    const updatedPackageJson = updatePackageJson(packageJson, [sourceDir]);
-                    fs.writeFileSync(
-                        packageJsonPath,
-                        JSON.stringify(updatedPackageJson, null, 2) + '\n'
-                    );
-                    console.log('✅ Added env-tool scripts to package.json');
-                } catch (error) {
-                    console.error('Failed to update package.json:', error);
-                    process.exitCode = 1;
-                    return;
-                }
-            }
-
-            // Create or update schema file
-            if (fs.existsSync(DEFAULT_SCHEMA_FILE) && !options.force) {
-                process.exitCode = 1;
-                console.warn(`${DEFAULT_SCHEMA_FILE} already exists. Use -f to overwrite if you want to start over.`)
-                return;
-            }
-
-            const vars = scanVars(sourceDir, options.git);
-            const out = initSchema(vars);
-            fs.writeFileSync(DEFAULT_SCHEMA_FILE, JSON.stringify(out, null, 4))
-            console.log(`✅ Created schema file at ${DEFAULT_SCHEMA_FILE}`);
-
-            // Print next steps
-            console.log(`
-                        Getting Started:
-
-                        1. Review and update the generated ${DEFAULT_SCHEMA_FILE} file:
-                        - Set appropriate default values
-                        - Add meaningful comments
-                        - Set required flags according to your needs
-
-                        2. Sync your .env file with the schema:
-                        npm run env:sync
-
-                        3. Validate your .env file:
-                        npm run env:validate
-                        `);
+        .action((dir: string | undefined, options: InitOptions) => {
+            initCommand(dir, options, DEFAULT_SCHEMA_FILE);
         })
 
     program.command('audit')
         .arguments('<dir>')
         .description('Audit project files for environment variables not in the schema')
         .option('--no-git', 'Disable Git tracking for file scanning')
-        .action((dir, options) => {
-            const vars = scanVars(dir, options.git);
-            const schema = loadSchema(DEFAULT_SCHEMA_FILE);
-            const issues = audit(vars, schema);
-            if (issues.length > 0) {
-                console.warn( issues.join('\n'))
-                process.exitCode = 1;
-            } else {
-                console.log('No issues found. All environment variables are in the schema.');
-            }
+        .action((dir: string, options: AuditOptions) => {
+            auditCommand(dir, options, DEFAULT_SCHEMA_FILE);
         })
 
     program.command('validate')
         .arguments('<envfile>')
         .description('Validate an env file against the schema')
-        .action((envfile) => {
-            const schema = loadSchema(DEFAULT_SCHEMA_FILE);
-            const envContent = fs.readFileSync(envfile);
-            const parsedEnv = parse(envContent);
-            const issues = validate(schema, parsedEnv);
-            if (issues.length > 0) {
-                const issueLabels = {
-                    no_value: 'has no value',
-                    not_defined: 'is not defined'
-                }
-                process.exitCode = 1;
-                console.warn(
-                    issues.map(iss => `${iss.key} ${issueLabels[iss.error]} in ${envfile}`).join('\n')
-                );
-            } else {
-                console.log(`${envfile} is valid.`);
-            }
+        .action((envfile: string) => {
+            validateCommand(envfile, DEFAULT_SCHEMA_FILE);
         });
 
     program.command('sync')
         .arguments('<envfile>')
         .description('Sync an env file with the schema')
-        .action((envfile) => {
-            const schema = loadSchema(DEFAULT_SCHEMA_FILE);
-
-            let envContent: Buffer;
-            try {
-                envContent = fs.readFileSync(envfile);
-                console.log(`Updating ${envfile} with schema values...`);
-            } catch(e: any) {
-                envContent = Buffer.from('');
-                console.log(`Creating new ${envfile} from schema...`);
-            }
-            const parsedEnv = parse(envContent);
-            const contents = syncEnvFile(schema, parsedEnv);
-            fs.writeFileSync(envfile, contents);
-            console.log(`${envfile} has been updated.`);
+        .action((envfile: string) => {
+            syncCommand(envfile, DEFAULT_SCHEMA_FILE);
         });
 
     return program;
